backend/requirements.txt
Django==4.2.7
mysqlclient==2.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-decouple==3.8
django-filter==23.3

django-browser-reload==1.10.0
